apps/backend/requirements.txt
uvicorn[standard]
fastapi
geoalchemy2
shapely
python-dotenv
psycopg2-binary
sqlalchemy
python-multipart
google-genai